+++ packages/color-helpers/types/global.d.ts
@@ -1,0 +1,5 @@
+export {};
+
+declare global {
+	type Color = [number, number, number];
+}