+++ plugins/postcss-hwb-function/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS HWB Function
 
+### Unreleased (minor)
+
+- Add: `@csstools/color-helpers` dependency for all color value transformations.
+
 ### 2.0.1 (January 28, 2023)
 
 - Improve `types` declaration in `package.json`